merge main

+++ newsfeed_project/src/main/java/com/sparta/newsfeed_project/domain/comment/service/CommentService.java
@@ -1,5 +1,6 @@
 package com.sparta.newsfeed_project.domain.comment.service;
 
+import com.sparta.newsfeed_project.auth.security.UserDetailsImpl;
 import com.sparta.newsfeed_project.domain.comment.dto.*;
 import com.sparta.newsfeed_project.domain.comment.entity.Comment;
 import com.sparta.newsfeed_project.domain.comment.repository.CommentRepository;
@@ -29,13 +30,13 @@ public class CommentService {
         Post post = postRepository.findById(post_id).
                 orElseThrow(()-> new NullPointerException("게시물을 찾을 수 없습니다"));
 
-        Comment comment = new Comment(commentSaveRequestDto.getContent(),commentSaveRequestDto.getPost(),commentSaveRequestDto.getUser());
+        Comment comment = new Comment(commentSaveRequestDto.getContent(),post,post.getUser());
         Comment savedComment = commentRepository.save(comment);
 
-        PostDto postDto = new PostDto(post_id,post.getTitle(),post.getContent(),post.getCreatedAt(),post.getModifiedAt());
+        PostDto postDto = new PostDto(post.getPostId(),post.getTitle(),post.getContent(),post.getCreatedAt(),post.getModifiedAt());
         return new CommentSaveResponseDto(
-                savedComment.getContent(),
                 postDto,
+                savedComment.getContent(),
                 savedComment.getCreatedAt(),
                 savedComment.getModifiedAt()
 
@@ -49,17 +50,18 @@ public class CommentService {
         Post post = postRepository.findById(post_id).
                 orElseThrow(()-> new NullPointerException("게시물을 찾을 수 없습니다"));
 
-        List<Comment> comments = commentRepository.findAll();
+        List<Comment> comments = post.getComments();
 
         List<CommentSimpleResponseDto> simpleDtoList = new ArrayList<>();
-        PostDto postDto =new PostDto(post.getPostId(),post.getTitle(),post.getContent(),post.getCreatedAt(),post.getModifiedAt());
         for(Comment comment : comments){
             CommentSimpleResponseDto dto = new CommentSimpleResponseDto(
+                    comment.getId(),
                     comment.getContent(),
+                    comment.getUser().getId(),
                     comment.getUser().getUsername(),
                     comment.getCreatedAt(),
-                    comment.getModifiedAt(),
-                    postDto);
+                    comment.getModifiedAt()
+                    );
             simpleDtoList.add(dto);
         }
         return simpleDtoList;
@@ -67,7 +69,8 @@ public class CommentService {
 
     //댓글 수정
     @Transactional
-    public CommentUpdateResponseDto updateComment(Long post_id,
+    public CommentUpdateResponseDto updateComment(UserDetailsImpl userDetails,
+                                                  Long post_id,
                                                   Long comment_Id,
                                                   CommentUpdateRequestDto commentUpdateRequestDto) {
 
@@ -79,9 +82,9 @@ public class CommentService {
                 .orElseThrow(() -> new NullPointerException("댓글을 찾을 수 없습니다."));
 
         //게시물의 작성자 or 댓글의 작성자인지 확인
-//        if (!comment.getUser().getId().equals() || post.getUser().getId().equals())(
-//                일치하지 않을 경우 예외처리!! 댓글 수정 권한이 없는 유저입니다.
-//                )
+        if (!comment.getUser().getId().equals(userDetails.getMyId()) || !post.getUser().getId().equals(userDetails.getMyId())){
+                throw new RuntimeException("댓글 수정은 댓글의 작성자 혹은 게시글의 작성자만 가능합니다.");
+                }
 
         comment.update(commentUpdateRequestDto.getContent());
         return new CommentUpdateResponseDto(
@@ -95,16 +98,18 @@ public class CommentService {
 
     //댓글 삭제
     @Transactional
-    public void deleteComment(Long post_id, Long comment_id) {
+    public void deleteComment(UserDetailsImpl userDetails,Long post_id, Long comment_id) {
         //게시글이 있는지 확인
-        postRepository.findById(post_id).orElseThrow(() -> new NullPointerException("게시글을 찾을 수 없습니다."));
+        Post post =postRepository.findById(post_id).orElseThrow(() -> new NullPointerException("게시글을 찾을 수 없습니다."));
         //댓글이 있는지 확인
         Comment comment = commentRepository.findById(comment_id)
                 .orElseThrow(() -> new NullPointerException("댓글을 찾을 수 없습니다."));
+
         //게시물의 작성자 or 댓글의 작성자인지 확인
-//        if (!comment.getUser().getId().equals() || post.getUser().getId().equals())(
-//                일치하지 않을 경우 예외처리!!! 댓글 삭제 권한이 없는 유저입니다.
-//                )
+        if (!comment.getUser().getId().equals(userDetails.getMyId()) || !post.getUser().getId().equals(userDetails.getMyId())){
+            throw new RuntimeException("댓글 삭제는 댓글의 작성자 혹은 게시글의 작성자만 가능합니다.");
+        }
+
 
         commentRepository.delete(comment);
